# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Techhype-API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/cards/": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/cards/{cardId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "active": {
                  "example": "any"
                },
                "designed": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                },
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "website": {
                  "example": "any"
                },
                "website2": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/workInfo/{profileId}": {
      "get": {
        "tags": [
          "WorkInfo"
        ],
        "description": "get Work info for a card profile.",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "WorkInfo"
        ],
        "description": "Add or update Work info for a card profile.",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/WorkInfo"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "WorkInfo"
        ],
        "description": "Delete Work info for a card profile.",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/workInfo/{workInfoId}": {
      "get": {
        "tags": [
          "WorkInfo"
        ],
        "description": "get Work info By using the WorkInfo ID.",
        "parameters": [
          {
            "name": "workInfoId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/socialMedia/{profileId}": {
      "get": {
        "tags": [
          "SocialMedia"
        ],
        "description": "get Social media for a card profile.",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "SocialMedia"
        ],
        "description": "Add or update a social media for a card profile.",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/SocialMedia"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "SocialMedia"
        ],
        "description": "Delete a social media for a card profile.",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/socialMedia/{socialMediaId}": {
      "get": {
        "tags": [
          "SocialMedia"
        ],
        "description": "get Social media by using the socialmedia ID.",
        "parameters": [
          {
            "name": "socialMediaId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  },
  "definitions": {
...
    "SocialMedia": {
      "type": "object",
      "properties": {
        "facebookLink": {
          "type": "string",
          "example": "www.facebook.com/Erik"
        },
        "linkedinLink": {
          "type": "string",
          "example": "www.linkedIn.com"
        },
        "snapLink": {
          "type": "string",
          "example": "www.snapchat.com"
        },
        "instagramLink": {
          "type": "string",
          "example": "www.instagram.com/akam"
        }
      },
      "required": [
        "facebookLink",
        "linkedinLink",
        "snapLink",
        "instagramLink"
      ]
    },
    "WorkInfo": {
      "type": "object",
      "properties": {
        "company": {
          "type": "string",
          "example": "Techhype AS"
        },
        "position": {
          "type": "string",
          "example": "CEO"
        },
        "workPhone": {
          "type": "string",
          "example": "9177128"
        },
        "workEmail": {
          "type": "string",
          "example": "glenn@techhype.no"
        }
      },
      "required": [
        "company",
        "position",
        "workPhone",
        "workEmail"
      ]
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}